+++ tests/Awesomite/ErrorDumper/Views/ViewHtmlTest.php
@@ -7,6 +7,8 @@ use Awesomite\ErrorDumper\Cloners\ClonedExceptionInterface;
 use Awesomite\ErrorDumper\Editors\EditorInterface;
 use Awesomite\ErrorDumper\Editors\Phpstorm;
 use Awesomite\ErrorDumper\TestBase;
+use Symfony\Component\Filesystem\Filesystem;
+use Symfony\Component\Finder\Finder;
 
 /**
  * @internal
@@ -62,5 +64,42 @@ class ViewHtmlTest extends TestBase
             array('<a href="http://localhost:8001">Test</a>'),
             array('Test'),
         );
+    }
+
+    public function testCache()
+    {
+        $cachePath = $this->getCachePath();
+        $this->assertInternalType('string', $cachePath);
+
+        $finder = new Finder();
+        $filesystem = new Filesystem();
+        $finder
+            ->ignoreDotFiles(true)
+            ->in($cachePath);
+        $filesystem->remove($finder);
+        $this->assertSame(0, count($finder));
+
+        $view = new ViewHtml();
+        $view->enableCaching($cachePath);
+        ob_start();
+        $view->display(new ClonedException(new \Exception()));
+        ob_end_clean();
+        $this->assertGreaterThan(0, count($finder));
+        $filesystem->remove($finder);
+
+        $view->disableCaching();
+        ob_start();
+        $view->display(new ClonedException(new \Exception()));
+        ob_end_clean();
+        $this->assertSame(0, count($finder));
+    }
+
+    private function getCachePath()
+    {
+        $exploded = explode(DIRECTORY_SEPARATOR . 'tests' . DIRECTORY_SEPARATOR, __DIR__);
+        array_pop($exploded);
+        $exploded = array_merge($exploded, array('tests', 'cache'));
+
+        return realpath(implode(DIRECTORY_SEPARATOR, $exploded));
     }
 }